+++ R/CALplot.R
@@ -13,9 +13,9 @@
 #' or \emph{uSe*} can be expected to be.
 #' @return This function returns a calibration plot
 #' @examples
-#' suit<-rbeta(100, 2, 2) #Generate suitability values
+#' suit<-rbeta(100, 2, 2) # Generate suitability values
 #' random<-runif(100)
-#' sp<-ifelse(random < suit,1 , 0) #Generate presence-absence data
+#' sp<-ifelse(random < suit,1 , 0) # Generate presence-absence data
 #' CALplot(cbind(suit, sp))
 #' @encoding UTF-8
 #' @references Jiménez-Valverde, A., Acevedo, P., Barbosa, A. M., Lobo, J. M. &
@@ -27,12 +27,11 @@
 CALplot <- function(mat, by = 0.1) {
     bins <- cut(mat[, 1], seq(0, 1, by), include.lowest = T, right = F)
     tableBins <- table(mat[, 2], bins)
-    prevalBins <- as.matrix(tableBins[2, ] / colSums(tableBins))
+    prevalBins <- as.matrix(tableBins[2, ]/colSums(tableBins))
     prevalBins.2 <- as.matrix(tapply(mat[, 1], bins, mean))
     colDots <- ifelse(colSums(tableBins) < 14, 21, 19)
-    plot(prevalBins.2, prevalBins, main = "Calibration plot", pch = colDots, ylab = "observed probability",
-        xlab = "predicted probability", xlim = c(0, 1), ylim = c(0, 1), yaxt = "n", cex.lab = 1.3,
-        cex = 1.1)
+    plot(prevalBins.2, prevalBins, main = "Calibration plot", pch = colDots, ylab = "observed probability", xlab = "predicted probability",
+        xlim = c(0, 1), ylim = c(0, 1), yaxt = "n", cex.lab = 1.3, cex = 1.1)
     axis(side = 2, las = 2, mgp = c(3, 0.75, 0))
     abline(a = 0, b = 1, col = "black", lty = 2)
 }