+++ R/HSgraph.R
@@ -16,9 +16,9 @@
 #' Jiménez-Valverde 2022).
 #' @return This function returns a multiple histogram.
 #' @examples
-#' suit<-rbeta(100, 2, 2) #Generate suitability values
+#' suit<-rbeta(100, 2, 2) # Generate suitability values
 #' random<-runif(100)
-#' sp<-ifelse(random < suit, 1 , 0) #Generate presence-absence data
+#' sp<-ifelse(random < suit, 1 , 0) # Generate presence-absence data
 #' HSgraph(cbind(suit, sp), breaks = 20, hist.total = TRUE)
 #' @encoding UTF-8
 #' @references Jiménez-Valverde, A. (2022). The uniform AUC: dealing with the
@@ -34,14 +34,12 @@ HSgraph <- function(mat, breaks = 10, hist.total = TRUE) {
     histpresence <- hist(mat[, 1][mat[, 2] == 1], breaks = breaks)
     histabsence <- hist(mat[, 1][mat[, 2] == 0], breaks = breaks)
     histtotal <- hist(mat[, 1], breaks = breaks)
-    plot(histpresence, col = rgb(0, 0, 1, 1 / 4), xlim = c(min(mat[, 1]),
-        max(mat[, 1])), ylim = c(0, max(histpresence$counts, histabsence$counts,
-        histtotal$counts)), border = rgb(0, 0, 1, 0), main = "",
-        xlab = "Suitability", ylab = "Frequency", las = 1)
-    plot(histabsence, col = rgb(1, 0, 0, 1 / 4), xlim = c(min(mat[, 1]),
-        max(mat[, 1])), border = rgb(1, 0, 0, 0), add = T)
+    plot(histpresence, col = rgb(0, 0, 1, 1/4), xlim = c(min(mat[, 1]), max(mat[, 1])), ylim = c(0, max(histpresence$counts,
+        histabsence$counts, histtotal$counts)), border = rgb(0, 0, 1, 0), main = "", xlab = "Suitability", ylab = "Frequency",
+        las = 1)
+    plot(histabsence, col = rgb(1, 0, 0, 1/4), xlim = c(min(mat[, 1]), max(mat[, 1])), border = rgb(1, 0, 0, 0),
+        add = T)
     if (hist.total == TRUE) {
-        lines(c(0, histtotal$breaks), c(0, histtotal$counts, 0), type = "s",
-        lwd = 2)
+        lines(c(0, histtotal$breaks), c(0, histtotal$counts, 0), type = "s", lwd = 2)
     }
 }